Combined runners into one for simplification, moved MemPatternsForPin and MemPatternsForNV into separate namespaces to avoid namespace conflicts. Renamed a few files to be more descriptive. Deleted old gsnv_test as its no longer needed.
