+++ CMakeLists.txt
@@ -14,18 +14,16 @@ add_library(gs_patterns_core SHARED
         gs_patterns.h
         gs_patterns_core.h
         gs_patterns_core.cpp
-        gsnv_patterns.h
-        gsnv_test.cpp
-#       nv_opcodes.h
 )
 
 add_executable( gs_patterns
-        gspin_patterns.cpp)
-
-add_executable(gsnv_test gsnv_test.cpp)
+        gs_patterns_main.cpp
+        gsnv_patterns.cpp
+        gspin_patterns.cpp
+        gspin_patterns.h
+        gsnv_patterns.h
+)
 
 target_link_libraries(gs_patterns gs_patterns_core)
-target_link_libraries(gsnv_test gs_patterns_core)
-
 
 set(CMAKE_CXX_STANDARD_LIBRARIES "-lm -lz ${CMAKE_CXX_STANDARD_LIBRARIES}")
