+++ nvbit_tracing/gsnv_trace/common.h
@@ -25,6 +25,8 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#pragma once
+
 #include <stdint.h>
 
 /* information collected in the instrumentation function and passed
